Backend/requirements.txt
# Core - API Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6

# Configuração e Ambiente
python-dotenv>=1.0.0

# Banco de Dados (opcional - se usar PostgreSQL)
psycopg[binary]>=3.2.0

# Segurança e Autenticação
python-jose[cryptography]>=3.3.0

# HTTP Client
httpx>=0.27.0

# Certificados e Criptografia
cryptography>=43.0.0
pyopenssl>=24.3.0

# Automação de Navegador
playwright>=1.46.0